initial distributions now have their own file
