+++ src/Cooked.hs
@@ -9,6 +9,7 @@ where
 
 import Cooked.Attack as X
 import Cooked.Currencies as X
+import Cooked.Distribution as X
 import qualified Cooked.Ltl as Ltl
 import Cooked.MockChain as X
 import Cooked.Output as X
