+++ src/Cooked/Wallet.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE NumericUnderscores #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeSynonymInstances #-}
 {-# OPTIONS_GHC -Wno-orphans #-}
@@ -20,24 +19,18 @@ module Cooked.Wallet
     walletAddress,
     walletSK,
     walletStakingSK,
-    initialDistribution,
-    InitialDistribution (..),
     Wallet,
     PrivateKey,
   )
 where
 
 import qualified Cardano.Crypto.Wallet as Cardano
-import Data.Default
 import Data.Function (on)
-import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import qualified Ledger.Address as Pl
 import qualified Ledger.CardanoWallet as Pl
 import qualified Ledger.Credential as Pl
 import qualified Ledger.Crypto as Pl
-import qualified Plutus.Script.Utils.Ada as Pl
-import qualified Plutus.Script.Utils.Value as Pl
 import Unsafe.Coerce
 
 -- * MockChain Wallets
@@ -126,51 +119,3 @@ walletStakingSK = fmap hackUnMockPrivateKey . Pl.mwStakeKey
     -- polymorphic because @MockPrivateKey@ is not exported either
     hackUnMockPrivateKey :: a -> PrivateKey
     hackUnMockPrivateKey x = let HACK y = unsafeCoerce x in y
-
--- * Initial distribution of funds
-
--- | Describes the initial distribution of UTxOs per wallet. This is important
--- since transaction validation must specify a /collateral/. Hence, wallets
--- must have more than one UTxO to begin with in order to execute a transaction
--- and have some collateral option. The @txCollateral@ is transferred to the
--- node operator in case the transaction fails to validate.
---
---  The following specifies a starting state where @wallet 1@ contains two
---  UTxOs, one with 42 Ada and one with 2 Ada and one "TOK" token; @wallet 2@
---  contains a single UTxO with 10 Ada and @wallet 3@ has 10 Ada and a
---  permanent value. See "Cooked.Currencies" for more information on quick and
---  permanent values.
---
---  > i0 = InitialDistribution $ M.fromList
---  >        [ (wallet 1 , [ Pl.lovelaveValueOf 42000000
---  >                      , Pl.lovelaceValueOf 2000000 <> quickValue "TOK" 1
---  >                      ]
---  >        , (wallet 2 , [Pl.lovelaveValueOf 10000000])
---  >        , (wallet 3 , [Pl.lovelaceValueOf 10000000 <> permanentValue "XYZ" 10])
---  >        ]
-newtype InitialDistribution = InitialDistribution
-  { unInitialDistribution :: Map Wallet [Pl.Value]
-  }
-  deriving (Eq, Show)
-
-instance Semigroup InitialDistribution where
-  (InitialDistribution i) <> (InitialDistribution j) =
-    InitialDistribution $ Map.unionWith (<>) i j
-
-instance Monoid InitialDistribution where
-  mempty = InitialDistribution Map.empty
-
--- | 10 UTxOs with 100 Ada each, for each of the 'knownWallets'.
-instance Default InitialDistribution where
-  def =
-    InitialDistribution $
-      Map.fromList $
-        zip knownWallets (repeat $ replicate 10 defLovelace)
-    where
-      defLovelace = Pl.lovelaceValueOf 100_000_000
-
-distributionFromList :: [(Wallet, [Pl.Value])] -> InitialDistribution
-distributionFromList = InitialDistribution . Map.fromList
-
-initialDistribution :: [(Wallet, [Pl.Value])] -> InitialDistribution
-initialDistribution = (def <>) . distributionFromList
